requirements.txt
aiogram==2.6.1
emoji-country-flag==1.2.1
SQLAlchemy==1.3.13
matplotlib==3.1.3